+++ modules/misc/Portal.php
@@ -78,6 +78,58 @@ do_action( 'misc/Portal.php|portal_alerts' );
 
 echo ErrorMessage( $note, 'note' );
 
+$portal_LO_options = array( 'save' => false, 'search' => false );
+
+$notes_LO_columns = array(
+	'PUBLISHED_DATE' => _( 'Date Posted' ),
+	'TITLE' => _( 'Title' ),
+	'CONTENT' => _( 'Note' ),
+	'FILE_ATTACHED' => _( 'File Attached' ),
+);
+
+$polls_LO_columns = array(
+	'PUBLISHED_DATE' => _( 'Date Posted' ),
+	'TITLE' => _( 'Title' ),
+	'OPTIONS' => _( 'Poll' ),
+);
+
+$events_LO_columns = array(
+	'DAY' => _( 'Day' ),
+	'SCHOOL_DATE' => _( 'Date' ),
+	'TITLE' => _( 'Event' ),
+	'DESCRIPTION' => _( 'Description' ),
+);
+
+if ( User( 'PROFILE') !== 'student'
+	|| SchoolInfo( 'SCHOOLS_NB' ) > 1 )
+{
+	// More than 1 school, display School column.
+	$notes_LO_columns['SCHOOL'] = _( 'School' );
+
+	$polls_LO_columns['SCHOOL'] = _( 'School' );
+
+	$events_LO_columns['SCHOOL'] = _( 'School' );
+}
+
+$assignments_LO_columns = array(
+	// 'DAY' => _( 'Day' ),
+	'DUE_DATE' => _( 'Due Date' ),
+	'ASSIGNMENT_TITLE' => _( 'Assignment' ),
+	// 'DESCRIPTION' => _( 'Notes' ),
+	'COURSE' => _( 'Course' ),
+	'STAFF_ID' => _( 'Teacher' ),
+	'SUBMITTED' => _( 'Submitted' ),
+);
+
+if ( User( 'PROFILE' ) === 'student'
+	|| User( 'PROFILE' ) === 'parent' )
+{
+	// Student or Parent, add Teacher & Submitted columns.
+	$assignments_LO_columns['STAFF_ID'] = _( 'Teacher' );
+	$assignments_LO_columns['SUBMITTED'] = _( 'Submitted' );
+}
+
+
 switch ( User( 'PROFILE' ) )
 {
 	case 'admin':
@@ -93,76 +145,109 @@ switch ( User( 'PROFILE' ) )
 		require_once 'ProgramFunctions/PortalPollsNotes.fnc.php';
 //FJ file attached to portal notes
 //FJ fix bug Portal Notes not displayed when pn.START_DATE IS NULL
-//        $notes_RET = DBGet(DBQuery("SELECT s.TITLE AS SCHOOL,date(pn.PUBLISHED_DATE) AS PUBLISHED_DATE,'<b>'||pn.TITLE||'</b>' AS TITLE,pn.CONTENT FROM PORTAL_NOTES pn,SCHOOLS s,STAFF st WHERE pn.SYEAR='".UserSyear()."' AND pn.START_DATE<=CURRENT_DATE AND (pn.END_DATE>=CURRENT_DATE OR pn.END_DATE IS NULL) AND st.STAFF_ID='".User('STAFF_ID')."' AND (st.SCHOOLS IS NULL OR position(','||pn.SCHOOL_ID||',' IN st.SCHOOLS)>0) AND (st.PROFILE_ID IS NULL AND position(',admin,' IN pn.PUBLISHED_PROFILES)>0 OR st.PROFILE_ID IS NOT NULL AND position(','||st.PROFILE_ID||',' IN pn.PUBLISHED_PROFILES)>0) AND s.ID=pn.SCHOOL_ID AND s.SYEAR=pn.SYEAR ORDER BY pn.SORT_ORDER,pn.PUBLISHED_DATE DESC"),array('PUBLISHED_DATE' => 'ProperDate','CONTENT' => '_formatContent'));
-		$notes_RET = DBGet(DBQuery("SELECT s.TITLE AS SCHOOL,date(pn.PUBLISHED_DATE) AS PUBLISHED_DATE,'<b>'||pn.TITLE||'</b>' AS TITLE,pn.CONTENT,pn.FILE_ATTACHED,pn.ID
+//        $notes_RET = DBGet( DBQuery( "SELECT s.TITLE AS SCHOOL,date(pn.PUBLISHED_DATE) AS PUBLISHED_DATE,'<b>'||pn.TITLE||'</b>' AS TITLE,pn.CONTENT FROM PORTAL_NOTES pn,SCHOOLS s,STAFF st WHERE pn.SYEAR='" . UserSyear() . "' AND pn.START_DATE<=CURRENT_DATE AND (pn.END_DATE>=CURRENT_DATE OR pn.END_DATE IS NULL) AND st.STAFF_ID='" . User( 'STAFF_ID' ) . "' AND (st.SCHOOLS IS NULL OR position(','||pn.SCHOOL_ID||',' IN st.SCHOOLS)>0) AND (st.PROFILE_ID IS NULL AND position(',admin,' IN pn.PUBLISHED_PROFILES)>0 OR st.PROFILE_ID IS NOT NULL AND position(','||st.PROFILE_ID||',' IN pn.PUBLISHED_PROFILES)>0) AND s.ID=pn.SCHOOL_ID AND s.SYEAR=pn.SYEAR ORDER BY pn.SORT_ORDER,pn.PUBLISHED_DATE DESC"),array('PUBLISHED_DATE' => 'ProperDate','CONTENT' => '_formatContent'));
+		$notes_RET = DBGet( DBQuery( "SELECT s.TITLE AS SCHOOL,date(pn.PUBLISHED_DATE) AS PUBLISHED_DATE,'<b>'||pn.TITLE||'</b>' AS TITLE,pn.CONTENT,pn.FILE_ATTACHED,pn.ID
 		FROM PORTAL_NOTES pn,SCHOOLS s,STAFF st
-		WHERE pn.SYEAR='".UserSyear()."'
+		WHERE pn.SYEAR='" . UserSyear() . "'
 		AND (pn.START_DATE<=CURRENT_DATE OR pn.START_DATE IS NULL)
 		AND (pn.END_DATE>=CURRENT_DATE OR pn.END_DATE IS NULL)
-		AND st.STAFF_ID='".User('STAFF_ID')."'
+		AND st.STAFF_ID='" . User( 'STAFF_ID' ) . "'
 		AND (st.SCHOOLS IS NULL OR position(','||pn.SCHOOL_ID||',' IN st.SCHOOLS)>0)
 		AND (st.PROFILE_ID IS NULL AND position(',admin,' IN pn.PUBLISHED_PROFILES)>0 OR st.PROFILE_ID IS NOT NULL AND position(','||st.PROFILE_ID||',' IN pn.PUBLISHED_PROFILES)>0)
 		AND s.ID=pn.SCHOOL_ID
 		AND s.SYEAR=pn.SYEAR
-		ORDER BY pn.SORT_ORDER,pn.PUBLISHED_DATE DESC"),array('PUBLISHED_DATE' => 'ProperDate', 'CONTENT' => 'makeTextarea', 'FILE_ATTACHED' => 'makeFileAttached'));
+		ORDER BY pn.SORT_ORDER,pn.PUBLISHED_DATE DESC" ),
+		array(
+			'PUBLISHED_DATE' => 'ProperDate',
+			'CONTENT' => 'makeTextarea',
+			'FILE_ATTACHED' => 'makeFileAttached',
+		) );
 
-		if (count($notes_RET))
+		if ( $notes_RET )
 		{
-			ListOutput($notes_RET,array('PUBLISHED_DATE' => _('Date Posted'),'TITLE' => _('Title'),'CONTENT' => _('Note'),'FILE_ATTACHED' => _('File Attached'),'SCHOOL' => _('School')),'Note','Notes',array(),array(),array('save'=>false,'search'=>false));
+			ListOutput(
+				$notes_RET,
+				$notes_LO_columns,
+				'Note',
+				'Notes',
+				array(),
+				array(),
+				$portal_LO_options
+			);
 		}
 
 		//FJ Portal Polls
-		$polls_RET = DBGet(DBQuery("SELECT s.TITLE AS SCHOOL,date(pp.PUBLISHED_DATE) AS PUBLISHED_DATE,'<b>'||pp.TITLE||'</b>' AS TITLE,'options' AS OPTIONS,pp.ID
+		$polls_RET = DBGet( DBQuery( "SELECT s.TITLE AS SCHOOL,date(pp.PUBLISHED_DATE) AS PUBLISHED_DATE,'<b>'||pp.TITLE||'</b>' AS TITLE,'options' AS OPTIONS,pp.ID
 		FROM PORTAL_POLLS pp,SCHOOLS s,STAFF st
-		WHERE pp.SYEAR='".UserSyear()."'
+		WHERE pp.SYEAR='" . UserSyear() . "'
 		AND (pp.START_DATE<=CURRENT_DATE OR pp.START_DATE IS NULL)
 		AND (pp.END_DATE>=CURRENT_DATE OR pp.END_DATE IS NULL)
-		AND st.STAFF_ID='".User('STAFF_ID')."'
+		AND st.STAFF_ID='" . User( 'STAFF_ID' ) . "'
 		AND (st.SCHOOLS IS NULL OR position(','||pp.SCHOOL_ID||',' IN st.SCHOOLS)>0)
 		AND (st.PROFILE_ID IS NULL AND position(',admin,' IN pp.PUBLISHED_PROFILES)>0 OR st.PROFILE_ID IS NOT NULL AND position(','||st.PROFILE_ID||',' IN pp.PUBLISHED_PROFILES)>0)
 		AND s.ID=pp.SCHOOL_ID AND s.SYEAR=pp.SYEAR
 		ORDER BY pp.SORT_ORDER,pp.PUBLISHED_DATE DESC"),array('PUBLISHED_DATE' => 'ProperDate','OPTIONS' => 'PortalPollsDisplay'));
 
-		if (count($polls_RET))
+		if ( $polls_RET )
 		{
-			ListOutput($polls_RET,array('PUBLISHED_DATE' => _('Date Posted'),'TITLE' => _('Title'),'OPTIONS' => _('Poll'),'SCHOOL' => _('School')),'Poll','Polls',array(),array(),array('save'=>false,'search'=>false));
+			ListOutput(
+				$polls_RET,
+				$polls_LO_columns,
+				'Poll',
+				'Polls',
+				array(),
+				array(),
+				$portal_LO_options
+			);
 		}
 
-//FJ add translation
-		$events_RET = DBGet(DBQuery("SELECT ce.ID,ce.TITLE,ce.DESCRIPTION,ce.SCHOOL_DATE AS SCHOOL_DATE,to_char(ce.SCHOOL_DATE,'Day') AS DAY,s.TITLE AS SCHOOL
+		$events_RET = DBGet( DBQuery( "SELECT ce.ID,ce.TITLE,ce.DESCRIPTION,ce.SCHOOL_DATE AS SCHOOL_DATE,to_char(ce.SCHOOL_DATE,'Day') AS DAY,s.TITLE AS SCHOOL
 		FROM CALENDAR_EVENTS ce,SCHOOLS s,STAFF st
 		WHERE ce.SCHOOL_DATE BETWEEN CURRENT_DATE
 		AND CURRENT_DATE+11
-		AND ce.SYEAR='".UserSyear()."'
-		AND st.STAFF_ID='".User('STAFF_ID')."'
+		AND ce.SYEAR='" . UserSyear() . "'
+		AND st.STAFF_ID='" . User( 'STAFF_ID' ) . "'
 		AND (st.SCHOOLS IS NULL OR position(','||ce.SCHOOL_ID||',' IN st.SCHOOLS)>0)
 		AND s.ID=ce.SCHOOL_ID
 		AND s.SYEAR=ce.SYEAR
-		ORDER BY ce.SCHOOL_DATE,s.TITLE"),array('SCHOOL_DATE' => 'ProperDate', 'DAY' => '_eventDay', 'DESCRIPTION' => 'makeTextarea'),array('SCHOOL_DATE'));
+		ORDER BY ce.SCHOOL_DATE,s.TITLE" ), array(
+			'SCHOOL_DATE' => 'ProperDate',
+			'DAY' => '_eventDay',
+			'DESCRIPTION' => 'makeTextarea'
+		) );
 
-		if (count($events_RET))
+		if ( $events_RET )
 		{
-			ListOutput($events_RET,array('DAY' => _('Day'),'SCHOOL_DATE' => _('Date'),'TITLE' => _('Event'),'DESCRIPTION' => _('Description'),'SCHOOL' => _('School')),'Day With Upcoming Events','Days With Upcoming Events',array(),array('SCHOOL_DATE'),array('save'=>false,'search'=>false));
+			ListOutput(
+				$events_RET,
+				$events_LO_columns,
+				'Day With Upcoming Events',
+				'Days With Upcoming Events',
+				array(),
+				array(),
+				$portal_LO_options
+			);
 		}
 
-        //RSSOutput(USER('PROFILE'));
+		//RSSOutput(USER('PROFILE'));
 
 		if (Preferences('HIDE_ALERTS')!='Y')
 		{
 		// warn if missing attendances
-		$categories_RET = DBGet(DBQuery("SELECT '0' AS ID,'Attendance' AS TITLE,0,NULL AS SORT_ORDER UNION SELECT ID,TITLE,1,SORT_ORDER FROM ATTENDANCE_CODE_CATEGORIES WHERE SYEAR='".UserSyear()."' AND SCHOOL_ID='".UserSchool()."' ORDER BY 3,SORT_ORDER"));
-		foreach ( (array) $categories_RET as $category)
+		$categories_RET = DBGet( DBQuery( "SELECT '0' AS ID,'Attendance' AS TITLE,0,NULL AS SORT_ORDER UNION SELECT ID,TITLE,1,SORT_ORDER FROM ATTENDANCE_CODE_CATEGORIES WHERE SYEAR='" . UserSyear() . "' AND SCHOOL_ID='".UserSchool()."' ORDER BY 3,SORT_ORDER"));
+
+		foreach ( (array) $categories_RET as $category )
 		{
-		//FJ days numbered
-		//FJ multiple school periods for a course period
+			// FJ days numbered.
+			// FJ multiple school periods for a course period.
 			if (SchoolInfo('NUMBER_DAYS_ROTATION') !== null)
 			{
 				$RET = DBGET(DBQuery("SELECT cp.COURSE_PERIOD_ID,s.TITLE AS SCHOOL,acc.SCHOOL_DATE,cp.TITLE
 				FROM ATTENDANCE_CALENDAR acc,COURSE_PERIODS cp,SCHOOL_PERIODS sp,SCHOOLS s,STAFF st, COURSE_PERIOD_SCHOOL_PERIODS cpsp
 				WHERE cp.COURSE_PERIOD_ID=cpsp.COURSE_PERIOD_ID
-				AND acc.SYEAR='".UserSyear()."'
+				AND acc.SYEAR='" . UserSyear() . "'
 				AND (acc.MINUTES IS NOT NULL AND acc.MINUTES>0)
-				AND st.STAFF_ID='".User('STAFF_ID')."'
+				AND st.STAFF_ID='" . User( 'STAFF_ID' ) . "'
 				AND (st.SCHOOLS IS NULL OR position(','||acc.SCHOOL_ID||',' IN st.SCHOOLS)>0)
 				AND cp.SCHOOL_ID=acc.SCHOOL_ID
 				AND cp.SYEAR=acc.SYEAR
@@ -181,14 +266,14 @@ switch ( User( 'PROFILE' ) )
 				AND position(',".$category['ID'].",' IN cp.DOES_ATTENDANCE)>0
 				AND s.ID=acc.SCHOOL_ID
 				AND s.SYEAR=acc.SYEAR
-				ORDER BY cp.TITLE,acc.SCHOOL_DATE"),array('SCHOOL_DATE' => 'ProperDate'),array('COURSE_PERIOD_ID'));
+				ORDER BY cp.TITLE,acc.SCHOOL_DATE"),array('SCHOOL_DATE' => 'ProperDate' ),array('COURSE_PERIOD_ID'));
 			} else {
 				$RET = DBGET(DBQuery("SELECT cp.COURSE_PERIOD_ID,s.TITLE AS SCHOOL,acc.SCHOOL_DATE,cp.TITLE
 				FROM ATTENDANCE_CALENDAR acc,COURSE_PERIODS cp,SCHOOL_PERIODS sp,SCHOOLS s,STAFF st, COURSE_PERIOD_SCHOOL_PERIODS cpsp
 				WHERE cp.COURSE_PERIOD_ID=cpsp.COURSE_PERIOD_ID
-				AND acc.SYEAR='".UserSyear()."'
+				AND acc.SYEAR='" . UserSyear() . "'
 				AND (acc.MINUTES IS NOT NULL AND acc.MINUTES>0)
-				AND st.STAFF_ID='".User('STAFF_ID')."'
+				AND st.STAFF_ID='" . User( 'STAFF_ID' ) . "'
 				AND (st.SCHOOLS IS NULL OR position(','||acc.SCHOOL_ID||',' IN st.SCHOOLS)>0)
 				AND cp.SCHOOL_ID=acc.SCHOOL_ID
 				AND cp.SYEAR=acc.SYEAR
@@ -201,42 +286,86 @@ switch ( User( 'PROFILE' ) )
 				AND position(',".$category['ID'].",' IN cp.DOES_ATTENDANCE)>0
 				AND s.ID=acc.SCHOOL_ID
 				AND s.SYEAR=acc.SYEAR
-				ORDER BY cp.TITLE,acc.SCHOOL_DATE"),array('SCHOOL_DATE' => 'ProperDate'),array('COURSE_PERIOD_ID'));
+				ORDER BY cp.TITLE,acc.SCHOOL_DATE"),array('SCHOOL_DATE' => 'ProperDate' ),array('COURSE_PERIOD_ID'));
 			}
 
-			if (count($RET))
+			if ( $RET )
 			{
-				echo ErrorMessage(array(_('Teachers have missing attendance data')), 'warning');
+				echo ErrorMessage( array( _( 'Teachers have missing attendance data' ) ), 'warning' );
 
-				ListOutput($RET,array('SCHOOL_DATE' => _('Date'),'TITLE' => _('Period').' '._('Days').' - '._('Short Name').' - '._('Teacher'),'SCHOOL' => _('School')),'Course Period with missing attendance data','Course Periods with missing attendance data',array(),array('COURSE_PERIOD_ID'),array('save'=>false,'search'=>false));
-//				echo '';
+				ListOutput(
+					$RET,
+					array(
+						'SCHOOL_DATE' => _( 'Date' ),
+						'TITLE' => _( 'Period' ) . ' ' . _( 'Days' ) . ' - ' . _( 'Short Name' ) . ' - ' . _( 'Teacher' ),
+						'SCHOOL' => _( 'School' ),
+					),
+					'Course Period with missing attendance data',
+					'Course Periods with missing attendance data',
+					array(),
+					array( 'COURSE_PERIOD_ID' ),
+					$portal_LO_options
+				);
 			}
 		}
 		}
 
-		if ( $RosarioModules['Food_Service'] && Preferences('HIDE_ALERTS')!='Y')
+		if ( $RosarioModules['Food_Service']
+			&& Preferences( 'HIDE_ALERTS' ) !== 'Y' )
 		{
 			$FS_config = ProgramConfig( 'food_service' );
 
 			// warn if negative food service balance
-			$staff = DBGet(DBQuery("SELECT (SELECT STATUS FROM FOOD_SERVICE_STAFF_ACCOUNTS WHERE STAFF_ID=s.STAFF_ID) AS STATUS,(SELECT BALANCE FROM FOOD_SERVICE_STAFF_ACCOUNTS WHERE STAFF_ID=s.STAFF_ID) AS BALANCE FROM STAFF s WHERE s.STAFF_ID='".User('STAFF_ID')."'"));
+			$staff = DBGet( DBQuery( "SELECT (SELECT STATUS
+					FROM FOOD_SERVICE_STAFF_ACCOUNTS
+					WHERE STAFF_ID=s.STAFF_ID) AS STATUS,
+				(SELECT BALANCE
+					FROM FOOD_SERVICE_STAFF_ACCOUNTS
+					WHERE STAFF_ID=s.STAFF_ID) AS BALANCE
+				FROM STAFF s
+				WHERE s.STAFF_ID='" . User( 'STAFF_ID' ) . "'" ) );
+
 			$staff = $staff[1];
-			if ( $staff['BALANCE'] && $staff['BALANCE']<0)
-				echo ErrorMessage(array(sprintf(_('You have a <b>negative</b> food service balance of <span style="color:red">%s</span>'),$staff['BALANCE'])), 'warning');
+
+			if ( $staff['BALANCE']
+				&& $staff['BALANCE'] < 0 )
+			{
+				echo ErrorMessage(
+					array( sprintf( _( 'You have a <b>negative</b> food service balance of <span style="color:red">%s</span>' ), $staff['BALANCE'] ) ),
+					'warning'
+				);
+			}
 
 			// warn if students with food service balances below minimum
 			$extra = array();
 			$extra['SELECT'] = ',fssa.STATUS,fsa.BALANCE';
 			$extra['FROM'] = ',FOOD_SERVICE_ACCOUNTS fsa,FOOD_SERVICE_STUDENT_ACCOUNTS fssa';
-			$extra['WHERE'] = ' AND fssa.STUDENT_ID=s.STUDENT_ID AND fsa.ACCOUNT_ID=fssa.ACCOUNT_ID AND fssa.STATUS IS NULL AND fsa.BALANCE<\''.$FS_config['FOOD_SERVICE_BALANCE_MINIMUM'][1]['VALUE'].'\'';
-			$_REQUEST['_search_all_schools'] = 'Y';
-			$RET = GetStuList($extra);
-			if (count($RET))
-			{
-			    echo ErrorMessage(array(sprintf(_('Some students have food service balances below %1.2f'),$FS_config['FOOD_SERVICE_BALANCE_MINIMUM'][1]['VALUE'])), 'warning');
+			$extra['WHERE'] = " AND fssa.STUDENT_ID=s.STUDENT_ID
+				AND fsa.ACCOUNT_ID=fssa.ACCOUNT_ID
+				AND fssa.STATUS IS NULL
+				AND fsa.BALANCE<'" . $FS_config['FOOD_SERVICE_BALANCE_MINIMUM'][1]['VALUE'] . "'";
 
-			    ListOutput($RET,array('FULL_NAME' => _('Student'),'GRADE_ID' => _('Grade Level'),'BALANCE' => _('Balance')),'Student','Students',array(),array(),array('save'=>false,'search'=>false));
-//			    echo '</p>';
+			$_REQUEST['_search_all_schools'] = 'Y';
+
+			$RET = GetStuList( $extra );
+
+			if ( $RET )
+			{
+				echo ErrorMessage(array(sprintf(_('Some students have food service balances below %1.2f' ),$FS_config['FOOD_SERVICE_BALANCE_MINIMUM'][1]['VALUE'])), 'warning');
+
+				ListOutput(
+					$RET,
+					array(
+						'FULL_NAME' => _( 'Student' ),
+						'GRADE_ID' => _( 'Grade Level' ),
+						'BALANCE' => _( 'Balance' ),
+					),
+					'Student',
+					'Students',
+					array(),
+					array(),
+					$portal_LO_options
+				);
 			}
 		}
 
@@ -247,13 +376,13 @@ switch ( User( 'PROFILE' ) )
 
 		require_once 'ProgramFunctions/PortalPollsNotes.fnc.php';
 //FJ fix bug Portal Notes not displayed when pn.START_DATE IS NULL
-//        $notes_RET = DBGet(DBQuery("SELECT s.TITLE AS SCHOOL,date(pn.PUBLISHED_DATE) AS PUBLISHED_DATE,'<b>'||pn.TITLE||'</b>' AS TITLE,pn.CONTENT FROM PORTAL_NOTES pn,SCHOOLS s,STAFF st WHERE pn.SYEAR='".UserSyear()."' AND pn.START_DATE<=CURRENT_DATE AND (pn.END_DATE>=CURRENT_DATE OR pn.END_DATE IS NULL) AND st.STAFF_ID='".User('STAFF_ID')."' AND (st.SCHOOLS IS NULL OR position(','||pn.SCHOOL_ID||',' IN st.SCHOOLS)>0) AND (st.PROFILE_ID IS NULL AND position(',teacher,' IN pn.PUBLISHED_PROFILES)>0 OR st.PROFILE_ID IS NOT NULL AND position(','||st.PROFILE_ID||',' IN pn.PUBLISHED_PROFILES)>0) AND s.ID=pn.SCHOOL_ID AND s.SYEAR=pn.SYEAR ORDER BY pn.SORT_ORDER,pn.PUBLISHED_DATE DESC"),array('PUBLISHED_DATE' => 'ProperDate','CONTENT' => '_formatContent'));
-        $notes_RET = DBGet(DBQuery("SELECT s.TITLE AS SCHOOL,date(pn.PUBLISHED_DATE) AS PUBLISHED_DATE,'<b>'||pn.TITLE||'</b>' AS TITLE,pn.CONTENT,pn.FILE_ATTACHED,pn.ID
+//        $notes_RET = DBGet( DBQuery( "SELECT s.TITLE AS SCHOOL,date(pn.PUBLISHED_DATE) AS PUBLISHED_DATE,'<b>'||pn.TITLE||'</b>' AS TITLE,pn.CONTENT FROM PORTAL_NOTES pn,SCHOOLS s,STAFF st WHERE pn.SYEAR='" . UserSyear() . "' AND pn.START_DATE<=CURRENT_DATE AND (pn.END_DATE>=CURRENT_DATE OR pn.END_DATE IS NULL) AND st.STAFF_ID='" . User( 'STAFF_ID' ) . "' AND (st.SCHOOLS IS NULL OR position(','||pn.SCHOOL_ID||',' IN st.SCHOOLS)>0) AND (st.PROFILE_ID IS NULL AND position(',teacher,' IN pn.PUBLISHED_PROFILES)>0 OR st.PROFILE_ID IS NOT NULL AND position(','||st.PROFILE_ID||',' IN pn.PUBLISHED_PROFILES)>0) AND s.ID=pn.SCHOOL_ID AND s.SYEAR=pn.SYEAR ORDER BY pn.SORT_ORDER,pn.PUBLISHED_DATE DESC"),array('PUBLISHED_DATE' => 'ProperDate','CONTENT' => '_formatContent'));
+		$notes_RET = DBGet( DBQuery( "SELECT s.TITLE AS SCHOOL,date(pn.PUBLISHED_DATE) AS PUBLISHED_DATE,'<b>'||pn.TITLE||'</b>' AS TITLE,pn.CONTENT,pn.FILE_ATTACHED,pn.ID
 		FROM PORTAL_NOTES pn,SCHOOLS s,STAFF st
-		WHERE pn.SYEAR='".UserSyear()."'
+		WHERE pn.SYEAR='" . UserSyear() . "'
 		AND (pn.START_DATE<=CURRENT_DATE OR pn.START_DATE IS NULL)
 		AND (pn.END_DATE>=CURRENT_DATE OR pn.END_DATE IS NULL)
-		AND st.STAFF_ID='".User('STAFF_ID')."'
+		AND st.STAFF_ID='" . User( 'STAFF_ID' ) . "'
 		AND (st.SCHOOLS IS NULL OR position(','||pn.SCHOOL_ID||',' IN st.SCHOOLS)>0)
 		AND (st.PROFILE_ID IS NULL AND position(',teacher,' IN pn.PUBLISHED_PROFILES)>0 OR st.PROFILE_ID IS NOT NULL
 		AND position(','||st.PROFILE_ID||',' IN pn.PUBLISHED_PROFILES)>0)
@@ -261,18 +390,26 @@ switch ( User( 'PROFILE' ) )
 		AND s.SYEAR=pn.SYEAR
 		ORDER BY pn.SORT_ORDER,pn.PUBLISHED_DATE DESC"),array('PUBLISHED_DATE' => 'ProperDate','CONTENT' => 'makeTextarea','FILE_ATTACHED' => 'makeFileAttached'));
 
-		if (count($notes_RET))
+		if ( $notes_RET )
 		{
-			ListOutput($notes_RET,array('PUBLISHED_DATE' => _('Date Posted'),'TITLE' => _('Title'),'CONTENT' => _('Note'),'FILE_ATTACHED' => _('File Attached'),'SCHOOL' => _('School')),'Note','Notes',array(),array(),array('save'=>false,'search'=>false));
+			ListOutput(
+				$notes_RET,
+				$notes_LO_columns,
+				'Note',
+				'Notes',
+				array(),
+				array(),
+				$portal_LO_options
+			);
 		}
 
-//FJ Portal Polls
-        $polls_RET = DBGet(DBQuery("SELECT s.TITLE AS SCHOOL,date(pp.PUBLISHED_DATE) AS PUBLISHED_DATE,'<b>'||pp.TITLE||'</b>' AS TITLE,'options' AS OPTIONS,pp.ID
+		// FJ Portal Polls.
+		$polls_RET = DBGet( DBQuery( "SELECT s.TITLE AS SCHOOL,date(pp.PUBLISHED_DATE) AS PUBLISHED_DATE,'<b>'||pp.TITLE||'</b>' AS TITLE,'options' AS OPTIONS,pp.ID
 		FROM PORTAL_POLLS pp,SCHOOLS s,STAFF st
-		WHERE pp.SYEAR='".UserSyear()."'
+		WHERE pp.SYEAR='" . UserSyear() . "'
 		AND (pp.START_DATE<=CURRENT_DATE OR pp.START_DATE IS NULL)
 		AND (pp.END_DATE>=CURRENT_DATE OR pp.END_DATE IS NULL)
-		AND st.STAFF_ID='".User('STAFF_ID')."'
+		AND st.STAFF_ID='" . User( 'STAFF_ID' ) . "'
 		AND (st.SCHOOLS IS NULL OR position(','||pp.SCHOOL_ID||',' IN st.SCHOOLS)>0)
 		AND (st.PROFILE_ID IS NULL AND position(',teacher,' IN pp.PUBLISHED_PROFILES)>0 OR st.PROFILE_ID IS NOT NULL
 		AND position(','||st.PROFILE_ID||',' IN pp.PUBLISHED_PROFILES)>0)
@@ -280,25 +417,44 @@ switch ( User( 'PROFILE' ) )
 		AND s.SYEAR=pp.SYEAR
 		ORDER BY pp.SORT_ORDER,pp.PUBLISHED_DATE DESC"),array('PUBLISHED_DATE' => 'ProperDate','OPTIONS' => 'PortalPollsDisplay'));
 
-		if (count($polls_RET))
+		if ( $polls_RET )
 		{
-			ListOutput($polls_RET,array('PUBLISHED_DATE' => _('Date Posted'),'TITLE' => _('Title'),'OPTIONS' => _('Poll'),'SCHOOL' => _('School')),'Poll','Polls',array(),array(),array('save'=>false,'search'=>false));
+			ListOutput(
+				$polls_RET,
+				$polls_LO_columns,
+				'Poll',
+				'Polls',
+				array(),
+				array(),
+				$portal_LO_options
+			);
 		}
 
-//FJ add translation
-		$events_RET = DBGet(DBQuery("SELECT ce.ID,ce.TITLE,ce.DESCRIPTION,ce.SCHOOL_DATE,to_char(ce.SCHOOL_DATE,'Day') AS DAY,s.TITLE AS SCHOOL
+		$events_RET = DBGet( DBQuery( "SELECT ce.ID,ce.TITLE,ce.DESCRIPTION,ce.SCHOOL_DATE,to_char(ce.SCHOOL_DATE,'Day') AS DAY,s.TITLE AS SCHOOL
 		FROM CALENDAR_EVENTS ce,SCHOOLS s
 		WHERE ce.SCHOOL_DATE BETWEEN CURRENT_DATE
 		AND CURRENT_DATE+11
-		AND ce.SYEAR='".UserSyear()."'
-		AND position(','||ce.SCHOOL_ID||',' IN (SELECT SCHOOLS FROM STAFF WHERE STAFF_ID='".User('STAFF_ID')."'))>0
+		AND ce.SYEAR='" . UserSyear() . "'
+		AND position(','||ce.SCHOOL_ID||',' IN (SELECT SCHOOLS FROM STAFF WHERE STAFF_ID='" . User( 'STAFF_ID' ) . "'))>0
 		AND s.ID=ce.SCHOOL_ID
 		AND s.SYEAR=ce.SYEAR
-		ORDER BY ce.SCHOOL_DATE,s.TITLE"),array('SCHOOL_DATE' => 'ProperDate', 'DAY' => '_eventDay', 'DESCRIPTION' => 'makeTextarea'),array('SCHOOL_DATE'));
+		ORDER BY ce.SCHOOL_DATE,s.TITLE" ), array(
+			'SCHOOL_DATE' => 'ProperDate',
+			'DAY' => '_eventDay',
+			'DESCRIPTION' => 'makeTextarea'
+		) );
 
-		if (count($events_RET))
+		if ( $events_RET )
 		{
-			ListOutput($events_RET,array('DAY' => _('Day'),'SCHOOL_DATE' => _('Date'),'TITLE' => _('Event'),'DESCRIPTION' => _('Description'),'SCHOOL' => _('School')),'Day With Upcoming Events','Days With Upcoming Events',array(),array('SCHOOL_DATE'),array('save'=>false,'search'=>false));
+			ListOutput(
+				$events_RET,
+				$events_LO_columns,
+				'Day With Upcoming Events',
+				'Days With Upcoming Events',
+				array(),
+				array(),
+				$portal_LO_options
+			);
 		}
 
 		// FJ Portal Assignments.
@@ -319,32 +475,25 @@ switch ( User( 'PROFILE' ) )
 			/*'DESCRIPTION' => 'makeTextarea',*/
 		) );
 
-		if ( count( $assignments_RET ) )
+		if ( $assignments_RET )
 		{
 			ListOutput(
 				$assignments_RET,
-				array(
-					/*'DAY' => _( 'Day' ),*/
-					'DUE_DATE' => _( 'Due Date' ),
-					'ASSIGNED_DATE' => _( 'Assigned Date' ),
-					'ASSIGNMENT_TITLE' => _( 'Assignment' ),
-					/*'DESCRIPTION' => _( 'Notes' ),*/
-					'COURSE' => _( 'Course' ),
-				),
+				$assignments_LO_columns,
 				'Upcoming Assignment',
 				'Upcoming Assignments',
 				array(),
 				array(),
-				array( 'save' => false, 'search' => false )
+				$portal_LO_options
 			);
 		}
 
-        //RSSOutput(USER('PROFILE'));
+		//RSSOutput(USER('PROFILE'));
 
 		if (Preferences('HIDE_ALERTS')!='Y')
 		{
 			// warn if missing attendances
-			$categories_RET = DBGet(DBQuery("SELECT '0' AS ID,'Attendance' AS TITLE,0,NULL AS SORT_ORDER UNION SELECT ID,TITLE,1,SORT_ORDER FROM ATTENDANCE_CODE_CATEGORIES WHERE SYEAR='".UserSyear()."' AND SCHOOL_ID='".UserSchool()."' ORDER BY 3,SORT_ORDER"));
+			$categories_RET = DBGet( DBQuery( "SELECT '0' AS ID,'Attendance' AS TITLE,0,NULL AS SORT_ORDER UNION SELECT ID,TITLE,1,SORT_ORDER FROM ATTENDANCE_CODE_CATEGORIES WHERE SYEAR='" . UserSyear() . "' AND SCHOOL_ID='".UserSchool()."' ORDER BY 3,SORT_ORDER"));
 			foreach ( (array) $categories_RET as $category)
 			{
 			//FJ days numbered
@@ -354,13 +503,13 @@ switch ( User( 'PROFILE' ) )
 					$RET = DBGET(DBQuery("SELECT cp.COURSE_PERIOD_ID,acc.SCHOOL_DATE,cp.TITLE
 					FROM ATTENDANCE_CALENDAR acc,COURSE_PERIODS cp,SCHOOL_PERIODS sp, COURSE_PERIOD_SCHOOL_PERIODS cpsp
 					WHERE cp.COURSE_PERIOD_ID=cpsp.COURSE_PERIOD_ID
-					AND acc.SYEAR='".UserSyear()."'
+					AND acc.SYEAR='" . UserSyear() . "'
 					AND (acc.MINUTES IS NOT NULL AND acc.MINUTES>0)
 					AND cp.SCHOOL_ID=acc.SCHOOL_ID
 					AND cp.SYEAR=acc.SYEAR
 					AND acc.SCHOOL_DATE<'".DBDate()."'
 					AND cp.CALENDAR_ID=acc.CALENDAR_ID
-					AND cp.TEACHER_ID='".User('STAFF_ID')."'
+					AND cp.TEACHER_ID='" . User( 'STAFF_ID' ) . "'
 					AND cp.MARKING_PERIOD_ID IN (SELECT MARKING_PERIOD_ID FROM SCHOOL_MARKING_PERIODS WHERE (MP='FY' OR MP='SEM' OR MP='QTR') AND SCHOOL_ID=acc.SCHOOL_ID AND acc.SCHOOL_DATE BETWEEN START_DATE AND END_DATE)
 					AND sp.PERIOD_ID=cpsp.PERIOD_ID
 					AND (sp.BLOCK IS NULL AND position(substring('MTWHFSU' FROM cast(
@@ -372,31 +521,41 @@ switch ( User( 'PROFILE' ) )
 					AS INT) FOR 1) IN cpsp.DAYS)>0 OR sp.BLOCK IS NOT NULL AND acc.BLOCK IS NOT NULL AND sp.BLOCK=acc.BLOCK)
 					AND NOT exists(SELECT '' FROM ATTENDANCE_COMPLETED ac WHERE ac.SCHOOL_DATE=acc.SCHOOL_DATE AND ac.STAFF_ID=cp.TEACHER_ID AND ac.PERIOD_ID=cpsp.PERIOD_ID AND TABLE_NAME='".$category['ID']."')
 					AND position(',".$category['ID'].",' IN cp.DOES_ATTENDANCE)>0
-					ORDER BY cp.TITLE,acc.SCHOOL_DATE"),array('SCHOOL_DATE' => 'ProperDate'),array('COURSE_PERIOD_ID'));
+					ORDER BY cp.TITLE,acc.SCHOOL_DATE"),array('SCHOOL_DATE' => 'ProperDate' ),array('COURSE_PERIOD_ID'));
 				} else {
 					$RET = DBGET(DBQuery("SELECT cp.COURSE_PERIOD_ID,acc.SCHOOL_DATE,cp.TITLE
 					FROM ATTENDANCE_CALENDAR acc,COURSE_PERIODS cp,SCHOOL_PERIODS sp, COURSE_PERIOD_SCHOOL_PERIODS cpsp
 					WHERE cp.COURSE_PERIOD_ID=cpsp.COURSE_PERIOD_ID
-					AND acc.SYEAR='".UserSyear()."'
+					AND acc.SYEAR='" . UserSyear() . "'
 					AND (acc.MINUTES IS NOT NULL AND acc.MINUTES>0)
 					AND cp.SCHOOL_ID=acc.SCHOOL_ID
 					AND cp.SYEAR=acc.SYEAR AND acc.SCHOOL_DATE<'".DBDate()."'
 					AND cp.CALENDAR_ID=acc.CALENDAR_ID
-					AND cp.TEACHER_ID='".User('STAFF_ID')."'
+					AND cp.TEACHER_ID='" . User( 'STAFF_ID' ) . "'
 					AND cp.MARKING_PERIOD_ID IN (SELECT MARKING_PERIOD_ID FROM SCHOOL_MARKING_PERIODS WHERE (MP='FY' OR MP='SEM' OR MP='QTR') AND SCHOOL_ID=acc.SCHOOL_ID AND acc.SCHOOL_DATE BETWEEN START_DATE AND END_DATE)
 					AND sp.PERIOD_ID=cpsp.PERIOD_ID
 					AND (sp.BLOCK IS NULL AND position(substring('UMTWHFS' FROM cast(extract(DOW FROM acc.SCHOOL_DATE) AS INT)+1 FOR 1) IN cpsp.DAYS)>0 OR sp.BLOCK IS NOT NULL AND acc.BLOCK IS NOT NULL AND sp.BLOCK=acc.BLOCK)
 					AND NOT exists(SELECT '' FROM ATTENDANCE_COMPLETED ac WHERE ac.SCHOOL_DATE=acc.SCHOOL_DATE AND ac.STAFF_ID=cp.TEACHER_ID AND ac.PERIOD_ID=cpsp.PERIOD_ID AND TABLE_NAME='".$category['ID']."')
 					AND position(',".$category['ID'].",' IN cp.DOES_ATTENDANCE)>0
-					ORDER BY cp.TITLE,acc.SCHOOL_DATE"),array('SCHOOL_DATE' => 'ProperDate'),array('COURSE_PERIOD_ID'));
+					ORDER BY cp.TITLE,acc.SCHOOL_DATE"),array('SCHOOL_DATE' => 'ProperDate' ),array('COURSE_PERIOD_ID'));
 				}
 
-				if (count($RET))
+				if ( $RET )
 				{
 					echo ErrorMessage(array(_('You have missing attendance data')), 'warning');
 
-					ListOutput($RET,array('SCHOOL_DATE' => _('Date'),'TITLE' => _('Period').' '._('Days').' - '._('Short Name').' - '._('Teacher')),'Course Period with missing attendance data','Course Periods with missing attendance data',array(),array('COURSE_PERIOD_ID'),array('save'=>false,'search'=>false));
-	//				echo '</p>';
+					ListOutput(
+						$RET,
+						array(
+							'SCHOOL_DATE' => _( 'Date' ),
+							'TITLE' => _( 'Period' ) . ' ' . _( 'Days' ) . ' - ' . _( 'Short Name' ) . ' - ' . _( 'Teacher' ),
+						),
+						'Course Period with missing attendance data',
+						'Course Periods with missing attendance data',
+						array(),
+						array( 'COURSE_PERIOD_ID' ),
+						$portal_LO_options
+					);
 				}
 			}
 		}
@@ -404,11 +563,11 @@ switch ( User( 'PROFILE' ) )
 		if ( $RosarioModules['Food_Service'] && Preferences('HIDE_ALERTS')!='Y')
 		{
 			// warn if negative food service balance
-			$staff = DBGet(DBQuery("SELECT (SELECT STATUS FROM FOOD_SERVICE_STAFF_ACCOUNTS WHERE STAFF_ID=s.STAFF_ID) AS STATUS,(SELECT BALANCE FROM FOOD_SERVICE_STAFF_ACCOUNTS WHERE STAFF_ID=s.STAFF_ID) AS BALANCE FROM STAFF s WHERE s.STAFF_ID='".User('STAFF_ID')."'"));
+			$staff = DBGet( DBQuery( "SELECT (SELECT STATUS FROM FOOD_SERVICE_STAFF_ACCOUNTS WHERE STAFF_ID=s.STAFF_ID) AS STATUS,(SELECT BALANCE FROM FOOD_SERVICE_STAFF_ACCOUNTS WHERE STAFF_ID=s.STAFF_ID) AS BALANCE FROM STAFF s WHERE s.STAFF_ID='" . User( 'STAFF_ID' ) . "'"));
 			$staff = $staff[1];
 
 			if ( $staff['BALANCE'] && $staff['BALANCE']<0)
-				echo ErrorMessage(array(sprintf(_('You have a <b>negative</b> food service balance of <span style="color:red">%s</span>'),$staff['BALANCE'])), 'warning');
+				echo ErrorMessage(array(sprintf(_('You have a <b>negative</b> food service balance of <span style="color:red">%s</span>' ),$staff['BALANCE'])), 'warning');
 		}
 
 		echo '<p>&nbsp;'._('Happy teaching...').'</p>';
@@ -418,32 +577,40 @@ switch ( User( 'PROFILE' ) )
 
 		require_once 'ProgramFunctions/PortalPollsNotes.fnc.php';
 //FJ fix bug Portal Notes not displayed when pn.START_DATE IS NULL
-//        $notes_RET = DBGet(DBQuery("SELECT s.TITLE AS SCHOOL,date(pn.PUBLISHED_DATE) AS PUBLISHED_DATE,pn.TITLE,pn.CONTENT FROM PORTAL_NOTES pn,SCHOOLS s,STAFF st WHERE pn.SYEAR='".UserSyear()."' AND pn.START_DATE<=CURRENT_DATE AND (pn.END_DATE>=CURRENT_DATE OR pn.END_DATE IS NULL) AND st.STAFF_ID='".User('STAFF_ID')."' AND pn.SCHOOL_ID IN (SELECT DISTINCT SCHOOL_ID FROM STUDENTS_JOIN_USERS sju, STUDENT_ENROLLMENT se WHERE sju.STAFF_ID='".User('STAFF_ID')."' AND se.SYEAR=pn.SYEAR AND se.STUDENT_ID=sju.STUDENT_ID AND se.START_DATE<=CURRENT_DATE AND (se.END_DATE>=CURRENT_DATE OR se.END_DATE IS NULL)) AND (st.SCHOOLS IS NULL OR position(','||pn.SCHOOL_ID||',' IN st.SCHOOLS)>0) AND (st.PROFILE_ID IS NULL AND position(',parent,' IN pn.PUBLISHED_PROFILES)>0 OR st.PROFILE_ID IS NOT NULL AND position(','||st.PROFILE_ID||',' IN pn.PUBLISHED_PROFILES)>0) AND s.ID=pn.SCHOOL_ID AND s.SYEAR=pn.SYEAR ORDER BY pn.SORT_ORDER,pn.PUBLISHED_DATE DESC"),array('PUBLISHED_DATE' => 'ProperDate','CONTENT' => '_formatContent'));
-		$notes_RET = DBGet(DBQuery("SELECT s.TITLE AS SCHOOL,date(pn.PUBLISHED_DATE) AS PUBLISHED_DATE,pn.TITLE,pn.CONTENT,pn.FILE_ATTACHED,pn.ID
+//        $notes_RET = DBGet( DBQuery( "SELECT s.TITLE AS SCHOOL,date(pn.PUBLISHED_DATE) AS PUBLISHED_DATE,pn.TITLE,pn.CONTENT FROM PORTAL_NOTES pn,SCHOOLS s,STAFF st WHERE pn.SYEAR='" . UserSyear() . "' AND pn.START_DATE<=CURRENT_DATE AND (pn.END_DATE>=CURRENT_DATE OR pn.END_DATE IS NULL) AND st.STAFF_ID='" . User( 'STAFF_ID' ) . "' AND pn.SCHOOL_ID IN (SELECT DISTINCT SCHOOL_ID FROM STUDENTS_JOIN_USERS sju, STUDENT_ENROLLMENT se WHERE sju.STAFF_ID='" . User( 'STAFF_ID' ) . "' AND se.SYEAR=pn.SYEAR AND se.STUDENT_ID=sju.STUDENT_ID AND se.START_DATE<=CURRENT_DATE AND (se.END_DATE>=CURRENT_DATE OR se.END_DATE IS NULL)) AND (st.SCHOOLS IS NULL OR position(','||pn.SCHOOL_ID||',' IN st.SCHOOLS)>0) AND (st.PROFILE_ID IS NULL AND position(',parent,' IN pn.PUBLISHED_PROFILES)>0 OR st.PROFILE_ID IS NOT NULL AND position(','||st.PROFILE_ID||',' IN pn.PUBLISHED_PROFILES)>0) AND s.ID=pn.SCHOOL_ID AND s.SYEAR=pn.SYEAR ORDER BY pn.SORT_ORDER,pn.PUBLISHED_DATE DESC"),array('PUBLISHED_DATE' => 'ProperDate','CONTENT' => '_formatContent'));
+		$notes_RET = DBGet( DBQuery( "SELECT s.TITLE AS SCHOOL,date(pn.PUBLISHED_DATE) AS PUBLISHED_DATE,pn.TITLE,pn.CONTENT,pn.FILE_ATTACHED,pn.ID
 		FROM PORTAL_NOTES pn,SCHOOLS s,STAFF st
-		WHERE pn.SYEAR='".UserSyear()."'
+		WHERE pn.SYEAR='" . UserSyear() . "'
 		AND (pn.START_DATE<=CURRENT_DATE OR pn.START_DATE IS NULL)
 		AND (pn.END_DATE>=CURRENT_DATE OR pn.END_DATE IS NULL)
-		AND st.STAFF_ID='".User('STAFF_ID')."'
-		AND pn.SCHOOL_ID IN (SELECT DISTINCT SCHOOL_ID FROM STUDENTS_JOIN_USERS sju, STUDENT_ENROLLMENT se WHERE sju.STAFF_ID='".User('STAFF_ID')."' AND se.SYEAR=pn.SYEAR AND se.STUDENT_ID=sju.STUDENT_ID AND se.START_DATE<=CURRENT_DATE AND (se.END_DATE>=CURRENT_DATE OR se.END_DATE IS NULL))
+		AND st.STAFF_ID='" . User( 'STAFF_ID' ) . "'
+		AND pn.SCHOOL_ID IN (SELECT DISTINCT SCHOOL_ID FROM STUDENTS_JOIN_USERS sju, STUDENT_ENROLLMENT se WHERE sju.STAFF_ID='" . User( 'STAFF_ID' ) . "' AND se.SYEAR=pn.SYEAR AND se.STUDENT_ID=sju.STUDENT_ID AND se.START_DATE<=CURRENT_DATE AND (se.END_DATE>=CURRENT_DATE OR se.END_DATE IS NULL))
 		AND (st.SCHOOLS IS NULL OR position(','||pn.SCHOOL_ID||',' IN st.SCHOOLS)>0)
 		AND (st.PROFILE_ID IS NULL AND position(',parent,' IN pn.PUBLISHED_PROFILES)>0 OR st.PROFILE_ID IS NOT NULL AND position(','||st.PROFILE_ID||',' IN pn.PUBLISHED_PROFILES)>0)
 		AND s.ID=pn.SCHOOL_ID
 		AND s.SYEAR=pn.SYEAR
 		ORDER BY pn.SORT_ORDER,pn.PUBLISHED_DATE DESC"),array('PUBLISHED_DATE' => 'ProperDate','CONTENT' => 'makeTextarea','FILE_ATTACHED' => 'makeFileAttached'));
 
-		if (count($notes_RET))
+		if ( $notes_RET )
 		{
-			ListOutput($notes_RET,array('PUBLISHED_DATE' => _('Date Posted'),'TITLE' => _('Title'),'CONTENT' => _('Note'),'FILE_ATTACHED' => _('File Attached'),'SCHOOL' => _('School')),'Note','Notes',array(),array(),array('save'=>false,'search'=>false));
+			ListOutput(
+				$notes_RET,
+				$notes_LO_columns,
+				'Note',
+				'Notes',
+				array(),
+				array(),
+				$portal_LO_options
+			);
 		}
 
-//FJ Portal Polls
-        $polls_RET = DBGet(DBQuery("SELECT s.TITLE AS SCHOOL,date(pp.PUBLISHED_DATE) AS PUBLISHED_DATE,'<b>'||pp.TITLE||'</b>' AS TITLE,'options' AS OPTIONS,pp.ID
+		// FJ Portal Polls.
+		$polls_RET = DBGet( DBQuery( "SELECT s.TITLE AS SCHOOL,date(pp.PUBLISHED_DATE) AS PUBLISHED_DATE,'<b>'||pp.TITLE||'</b>' AS TITLE,'options' AS OPTIONS,pp.ID
 		FROM PORTAL_POLLS pp,SCHOOLS s,STAFF st
-		WHERE pp.SYEAR='".UserSyear()."'
+		WHERE pp.SYEAR='" . UserSyear() . "'
 		AND (pp.START_DATE<=CURRENT_DATE OR pp.START_DATE IS NULL)
 		AND (pp.END_DATE>=CURRENT_DATE OR pp.END_DATE IS NULL)
-		AND st.STAFF_ID='".User('STAFF_ID')."'
+		AND st.STAFF_ID='" . User( 'STAFF_ID' ) . "'
 		AND (st.SCHOOLS IS NULL OR position(','||pp.SCHOOL_ID||',' IN st.SCHOOLS)>0)
 		AND (st.PROFILE_ID IS NULL AND position(',parent,' IN pp.PUBLISHED_PROFILES)>0 OR st.PROFILE_ID IS NOT NULL
 		AND position(','||st.PROFILE_ID||',' IN pp.PUBLISHED_PROFILES)>0)
@@ -451,24 +618,43 @@ switch ( User( 'PROFILE' ) )
 		AND s.SYEAR=pp.SYEAR
 		ORDER BY pp.SORT_ORDER,pp.PUBLISHED_DATE DESC"),array('PUBLISHED_DATE' => 'ProperDate','OPTIONS' => 'PortalPollsDisplay'));
 
-		if (count($polls_RET))
+		if ( $polls_RET )
 		{
-			ListOutput($polls_RET,array('PUBLISHED_DATE' => _('Date Posted'),'TITLE' => _('Title'),'OPTIONS' => _('Poll'),'SCHOOL' => _('School')),'Poll','Polls',array(),array(),array('save'=>false,'search'=>false));
+			ListOutput(
+				$polls_RET,
+				$polls_LO_columns,
+				'Poll',
+				'Polls',
+				array(),
+				array(),
+				$portal_LO_options
+			);
 		}
 
-//FJ add translation
-		$events_RET = DBGet(DBQuery("SELECT ce.ID,ce.TITLE,ce.SCHOOL_DATE,to_char(ce.SCHOOL_DATE,'Day') AS DAY,ce.DESCRIPTION,s.TITLE AS SCHOOL
+		$events_RET = DBGet( DBQuery( "SELECT ce.ID,ce.TITLE,ce.SCHOOL_DATE,to_char(ce.SCHOOL_DATE,'Day') AS DAY,ce.DESCRIPTION,s.TITLE AS SCHOOL
 		FROM CALENDAR_EVENTS ce,SCHOOLS s
 		WHERE ce.SCHOOL_DATE BETWEEN CURRENT_DATE AND CURRENT_DATE+11
-		AND ce.SYEAR='".UserSyear()."'
-		AND ce.SCHOOL_ID IN (SELECT DISTINCT SCHOOL_ID FROM STUDENTS_JOIN_USERS sju, STUDENT_ENROLLMENT se WHERE sju.STAFF_ID='".User('STAFF_ID')."' AND se.SYEAR=ce.SYEAR AND se.STUDENT_ID=sju.STUDENT_ID AND se.START_DATE<=CURRENT_DATE AND (se.END_DATE>=CURRENT_DATE OR se.END_DATE IS NULL))
+		AND ce.SYEAR='" . UserSyear() . "'
+		AND ce.SCHOOL_ID IN (SELECT DISTINCT SCHOOL_ID FROM STUDENTS_JOIN_USERS sju, STUDENT_ENROLLMENT se WHERE sju.STAFF_ID='" . User( 'STAFF_ID' ) . "' AND se.SYEAR=ce.SYEAR AND se.STUDENT_ID=sju.STUDENT_ID AND se.START_DATE<=CURRENT_DATE AND (se.END_DATE>=CURRENT_DATE OR se.END_DATE IS NULL))
 		AND s.ID=ce.SCHOOL_ID
 		AND s.SYEAR=ce.SYEAR
-		ORDER BY ce.SCHOOL_DATE,s.TITLE"),array('SCHOOL_DATE' => 'ProperDate', 'DAY' => '_eventDay', 'DESCRIPTION' => 'makeTextarea'),array('SCHOOL_DATE'));
+		ORDER BY ce.SCHOOL_DATE,s.TITLE" ), array(
+			'SCHOOL_DATE' => 'ProperDate',
+			'DAY' => '_eventDay',
+			'DESCRIPTION' => 'makeTextarea'
+		) );
 
-		if (count($events_RET))
+		if ( $events_RET )
 		{
-			ListOutput($events_RET,array('DAY' => _('Day'),'SCHOOL_DATE' => _('Date'),'TITLE' => _('Event'),'DESCRIPTION' => _('Description'),'SCHOOL' => _('School')),'Day With Upcoming Events','Days With Upcoming Events',array(),array('SCHOOL_DATE'),array('save'=>false,'search'=>false));
+			ListOutput(
+				$events_RET,
+				$events_LO_columns,
+				'Day With Upcoming Events',
+				'Days With Upcoming Events',
+				array(),
+				array(),
+				$portal_LO_options
+			);
 		}
 
 		// FJ Portal Assignments.
@@ -502,55 +688,69 @@ switch ( User( 'PROFILE' ) )
 				'ASSIGNMENT_TITLE' => 'MakeAssignmentTitle',
 			) );
 
-			if ( count( $assignments_RET ) )
+			if ( $assignments_RET )
 			{
 				ListOutput(
 					$assignments_RET,
-					array(
-						/*'DAY' => _( 'Day' ),*/
-						'DUE_DATE' => _( 'Due Date' ),
-						'ASSIGNMENT_TITLE' => _( 'Assignment' ),
-						/*'DESCRIPTION' => _( 'Notes' ),*/
-						'COURSE' => _( 'Course' ),
-						'STAFF_ID' => _( 'Teacher' ),
-						'SUBMITTED' => _( 'Submitted' ),
-					),
+					$assignments_LO_columns,
 					'Upcoming Assignment',
 					'Upcoming Assignments',
 					array(),
 					array(),
-					array( 'save' => false, 'search' => false )
+					$portal_LO_options
 				);
 			}
 		}
 
-        //RSSOutput(USER('PROFILE'));
+		//RSSOutput(USER('PROFILE'));
 
-		if ( $RosarioModules['Food_Service'] && Preferences('HIDE_ALERTS')!='Y')
+		if ( $RosarioModules['Food_Service']
+			&& Preferences( 'HIDE_ALERTS' ) !== 'Y' )
 		{
 			$FS_config = ProgramConfig( 'food_service' );
 
-			// warn if students with low food service balances
-			$extra['SELECT'] = ',fssa.STATUS,fsa.ACCOUNT_ID,fsa.BALANCE AS BALANCE,'.$FS_config['FOOD_SERVICE_BALANCE_TARGET'][1]['VALUE'].'-fsa.BALANCE AS DEPOSIT';
+			// Warn if students with low food service balances.
+			$extra['SELECT'] = ',fssa.STATUS,fsa.ACCOUNT_ID,fsa.BALANCE AS BALANCE,' .
+				$FS_config['FOOD_SERVICE_BALANCE_TARGET'][1]['VALUE'] . '-fsa.BALANCE AS DEPOSIT';
 			$extra['FROM'] = ',FOOD_SERVICE_ACCOUNTS fsa,FOOD_SERVICE_STUDENT_ACCOUNTS fssa';
-			$extra['WHERE'] = ' AND fssa.STUDENT_ID=s.STUDENT_ID AND fsa.ACCOUNT_ID=fssa.ACCOUNT_ID AND fssa.STATUS IS NULL AND fsa.BALANCE<\''.$FS_config['FOOD_SERVICE_BALANCE_WARNING'][1]['VALUE'].'\'';
-			$extra['ASSOCIATED'] = User('STAFF_ID');
+			$extra['WHERE'] = " AND fssa.STUDENT_ID=s.STUDENT_ID
+				AND fsa.ACCOUNT_ID=fssa.ACCOUNT_ID
+				AND fssa.STATUS IS NULL
+				AND fsa.BALANCE<'" . $FS_config['FOOD_SERVICE_BALANCE_WARNING'][1]['VALUE'] . "'";
+			$extra['ASSOCIATED'] = User( 'STAFF_ID' );
 
-			$RET = GetStuList($extra);
+			$RET = GetStuList( $extra );
 
-			if (count($RET))
+			if ( $RET )
 			{
-				echo ErrorMessage(array(sprintf(_('You have students with food service balance below %1.2f - please deposit at least the Minimum Deposit into you children\'s accounts.'),$FS_config['FOOD_SERVICE_BALANCE_WARNING'][1]['VALUE'])), 'warning');
+				echo ErrorMessage(
+					array( sprintf( _( 'You have students with food service balance below %1.2f - please deposit at least the Minimum Deposit into you children\'s accounts.' ), $FS_config['FOOD_SERVICE_BALANCE_WARNING'][1]['VALUE'] ) ),
+					'warning'
+				);
 
-				ListOutput($RET,array('FULL_NAME' => _('Student'),'GRADE_ID' => _('Grade Level'),'ACCOUNT_ID' => _('Account ID'),'BALANCE' => _('Balance'),'DEPOSIT' => _('Minimum Deposit')),'Student','Students',array(),array(),array('save'=>false,'search'=>false));
+				ListOutput(
+					$RET,
+					array(
+						'FULL_NAME' => _( 'Student' ),
+						'GRADE_ID' => _( 'Grade Level' ),
+						'ACCOUNT_ID' => _( 'Account ID' ),
+						'BALANCE' => _( 'Balance' ),
+						'DEPOSIT' => _( 'Minimum Deposit' ),
+					),
+					'Student',
+					'Students',
+					array(),
+					array(),
+					$portal_LO_options
+				);
 			}
 
-			// warn if negative food service balance
-			$staff = DBGet(DBQuery("SELECT (SELECT STATUS FROM FOOD_SERVICE_STAFF_ACCOUNTS WHERE STAFF_ID=s.STAFF_ID) AS STATUS,(SELECT BALANCE FROM FOOD_SERVICE_STAFF_ACCOUNTS WHERE STAFF_ID=s.STAFF_ID) AS BALANCE FROM STAFF s WHERE s.STAFF_ID='".User('STAFF_ID')."'"));
+			// Warn if negative food service balance.
+			$staff = DBGet( DBQuery( "SELECT (SELECT STATUS FROM FOOD_SERVICE_STAFF_ACCOUNTS WHERE STAFF_ID=s.STAFF_ID) AS STATUS,(SELECT BALANCE FROM FOOD_SERVICE_STAFF_ACCOUNTS WHERE STAFF_ID=s.STAFF_ID) AS BALANCE FROM STAFF s WHERE s.STAFF_ID='" . User( 'STAFF_ID' ) . "'"));
 			$staff = $staff[1];
 
 			if ( $staff['BALANCE'] && $staff['BALANCE']<0)
-				echo ErrorMessage(array(sprintf(_('You have a <b>negative</b> food service balance of <span style="color:red">%s</span>'),Currency($staff['BALANCE']))), 'warning');
+				echo ErrorMessage(array(sprintf(_('You have a <b>negative</b> food service balance of <span style="color:red">%s</span>' ),Currency($staff['BALANCE']))), 'warning');
 		}
 
 		echo '<p>&nbsp;'._('Happy parenting...').'</p>';
@@ -560,10 +760,10 @@ switch ( User( 'PROFILE' ) )
 
 		require_once 'ProgramFunctions/PortalPollsNotes.fnc.php';
 //FJ fix bug Portal Notes not displayed when pn.START_DATE IS NULL
-//        $notes_RET = DBGet(DBQuery("SELECT s.TITLE AS SCHOOL,date(pn.PUBLISHED_DATE) AS PUBLISHED_DATE,pn.TITLE,pn.CONTENT FROM PORTAL_NOTES pn,SCHOOLS s WHERE pn.SYEAR='".UserSyear()."' AND pn.START_DATE<=CURRENT_DATE AND (pn.END_DATE>=CURRENT_DATE OR pn.END_DATE IS NULL) AND pn.SCHOOL_ID='".UserSchool()."' AND  position(',0,' IN pn.PUBLISHED_PROFILES)>0 AND s.ID=pn.SCHOOL_ID AND s.SYEAR=pn.SYEAR ORDER BY pn.SORT_ORDER,pn.PUBLISHED_DATE DESC"),array('PUBLISHED_DATE' => 'ProperDate','CONTENT' => '_formatContent'));
-        $notes_RET = DBGet(DBQuery("SELECT s.TITLE AS SCHOOL,date(pn.PUBLISHED_DATE) AS PUBLISHED_DATE,pn.TITLE,pn.CONTENT,pn.FILE_ATTACHED,pn.ID
+//        $notes_RET = DBGet( DBQuery( "SELECT s.TITLE AS SCHOOL,date(pn.PUBLISHED_DATE) AS PUBLISHED_DATE,pn.TITLE,pn.CONTENT FROM PORTAL_NOTES pn,SCHOOLS s WHERE pn.SYEAR='" . UserSyear() . "' AND pn.START_DATE<=CURRENT_DATE AND (pn.END_DATE>=CURRENT_DATE OR pn.END_DATE IS NULL) AND pn.SCHOOL_ID='".UserSchool()."' AND  position(',0,' IN pn.PUBLISHED_PROFILES)>0 AND s.ID=pn.SCHOOL_ID AND s.SYEAR=pn.SYEAR ORDER BY pn.SORT_ORDER,pn.PUBLISHED_DATE DESC"),array('PUBLISHED_DATE' => 'ProperDate','CONTENT' => '_formatContent'));
+		$notes_RET = DBGet( DBQuery( "SELECT s.TITLE AS SCHOOL,date(pn.PUBLISHED_DATE) AS PUBLISHED_DATE,pn.TITLE,pn.CONTENT,pn.FILE_ATTACHED,pn.ID
 		FROM PORTAL_NOTES pn,SCHOOLS s
-		WHERE pn.SYEAR='".UserSyear()."'
+		WHERE pn.SYEAR='" . UserSyear() . "'
 		AND (pn.START_DATE<=CURRENT_DATE OR pn.START_DATE IS NULL)
 		AND (pn.END_DATE>=CURRENT_DATE OR pn.END_DATE IS NULL)
 		AND pn.SCHOOL_ID='".UserSchool()."'
@@ -572,39 +772,67 @@ switch ( User( 'PROFILE' ) )
 		AND s.SYEAR=pn.SYEAR
 		ORDER BY pn.SORT_ORDER,pn.PUBLISHED_DATE DESC"),array('PUBLISHED_DATE' => 'ProperDate','CONTENT' => 'makeTextarea','FILE_ATTACHED' => 'makeFileAttached'));
 
-		if (count($notes_RET))
+		if ( $notes_RET )
 		{
-			ListOutput($notes_RET,array('PUBLISHED_DATE' => _('Date Posted'),'TITLE' => _('Title'),'CONTENT' => _('Note'),'FILE_ATTACHED' => _('File Attached')),'Note','Notes',array(),array(),array('save'=>false,'search'=>false));
+			ListOutput(
+				$notes_RET,
+				$notes_LO_columns,
+				'Note',
+				'Notes',
+				array(),
+				array(),
+				$portal_LO_options
+			);
 		}
 
-//FJ Portal Polls
-		//FJ Portal Polls add students teacher
-        $polls_RET = DBGet(DBQuery("SELECT s.TITLE AS SCHOOL,date(pp.PUBLISHED_DATE) AS PUBLISHED_DATE,pp.TITLE,'options' AS OPTIONS,pp.ID
+		// FJ Portal Polls.
+		// FJ Portal Polls add students teacher.
+		$polls_RET = DBGet( DBQuery( "SELECT s.TITLE AS SCHOOL,date(pp.PUBLISHED_DATE) AS PUBLISHED_DATE,pp.TITLE,'options' AS OPTIONS,pp.ID
 		FROM PORTAL_POLLS pp,SCHOOLS s
-		WHERE pp.SYEAR='".UserSyear()."'
+		WHERE pp.SYEAR='" . UserSyear() . "'
 		AND (pp.START_DATE<=CURRENT_DATE OR pp.START_DATE IS NULL)
 		AND (pp.END_DATE>=CURRENT_DATE OR pp.END_DATE IS NULL)
 		AND pp.SCHOOL_ID='".UserSchool()."'
 		AND position(',0,' IN pp.PUBLISHED_PROFILES)>0
 		AND s.ID=pp.SCHOOL_ID
 		AND s.SYEAR=pp.SYEAR
-		AND (pp.STUDENTS_TEACHER_ID IS NULL OR pp.STUDENTS_TEACHER_ID IN (SELECT cp.TEACHER_ID FROM SCHEDULE sch, COURSE_PERIODS cp WHERE sch.SYEAR='".UserSyear()."' AND sch.SCHOOL_ID='".UserSchool()."' AND sch.STUDENT_ID='".UserStudentID()."' AND sch.COURSE_PERIOD_ID=cp.COURSE_PERIOD_ID))
+		AND (pp.STUDENTS_TEACHER_ID IS NULL OR pp.STUDENTS_TEACHER_ID IN (SELECT cp.TEACHER_ID FROM SCHEDULE sch, COURSE_PERIODS cp WHERE sch.SYEAR='" . UserSyear() . "' AND sch.SCHOOL_ID='".UserSchool()."' AND sch.STUDENT_ID='".UserStudentID()."' AND sch.COURSE_PERIOD_ID=cp.COURSE_PERIOD_ID))
 		ORDER BY pp.SORT_ORDER,pp.PUBLISHED_DATE DESC"),array('PUBLISHED_DATE' => 'ProperDate','OPTIONS' => 'PortalPollsDisplay'));
 
-		if (count($polls_RET))
+		if ( $polls_RET )
 		{
-			ListOutput($polls_RET,array('PUBLISHED_DATE' => _('Date Posted'),'TITLE' => _('Title'),'OPTIONS' => _('Poll'),'SCHOOL' => _('School')),'Poll','Polls',array(),array(),array('save'=>false,'search'=>false));
+			ListOutput(
+				$polls_RET,
+				$polls_LO_columns,
+				'Poll',
+				'Polls',
+				array(),
+				array(),
+				$portal_LO_options
+			);
 		}
 
-		$events_RET = DBGet(DBQuery("SELECT ID,TITLE,SCHOOL_DATE,to_char(SCHOOL_DATE,'Day') AS DAY,DESCRIPTION
+		$events_RET = DBGet( DBQuery( "SELECT ID,TITLE,SCHOOL_DATE,to_char(SCHOOL_DATE,'Day') AS DAY,DESCRIPTION
 		FROM CALENDAR_EVENTS
 		WHERE SCHOOL_DATE BETWEEN CURRENT_DATE AND CURRENT_DATE+11
-		AND SYEAR='".UserSyear()."'
-		AND SCHOOL_ID='".UserSchool()."'"),array('SCHOOL_DATE' => 'ProperDate', 'DAY' => '_eventDay', 'DESCRIPTION' => 'makeTextarea'),array('SCHOOL_DATE'));
+		AND SYEAR='" . UserSyear() . "'
+		AND SCHOOL_ID='" . UserSchool() . "'" ), array(
+			'SCHOOL_DATE' => 'ProperDate',
+			'DAY' => '_eventDay',
+			'DESCRIPTION' => 'makeTextarea'
+		) );
 
-		if (count($events_RET))
+		if ( $events_RET )
 		{
-			ListOutput($events_RET,array('DAY' => _('Day'),'SCHOOL_DATE' => _('Date'),'TITLE' => _('Event'),'DESCRIPTION' => _('Description')),'Day With Upcoming Events','Days With Upcoming Events',array(),array('SCHOOL_DATE'),array('save'=>false,'search'=>false));
+			ListOutput(
+				$events_RET,
+				$events_LO_columns,
+				'Day With Upcoming Events',
+				'Days With Upcoming Events',
+				array(),
+				array(),
+				$portal_LO_options
+			);
 		}
 
 		// FJ Portal Assignments.
@@ -638,29 +866,21 @@ switch ( User( 'PROFILE' ) )
 				'ASSIGNMENT_TITLE' => 'MakeAssignmentTitle',
 			) );
 
-			if ( count( $assignments_RET ) )
+			if ( $assignments_RET )
 			{
 				ListOutput(
 					$assignments_RET,
-					array(
-						/*'DAY' => _( 'Day' ),*/
-						'DUE_DATE' => _( 'Due Date' ),
-						'ASSIGNMENT_TITLE' => _( 'Assignment' ),
-						/*'DESCRIPTION' => _( 'Notes' ),*/
-						'COURSE' => _( 'Course' ),
-						'STAFF_ID' => _( 'Teacher' ),
-						'SUBMITTED' => _( 'Submitted' ),
-					),
+					$assignments_LO_columns,
 					'Upcoming Assignment',
 					'Upcoming Assignments',
 					array(),
 					array(),
-					array( 'save' => false, 'search' => false )
+					$portal_LO_options
 				);
 			}
 		}
 
-        //RSSOutput(USER('PROFILE'));
+		//RSSOutput(USER('PROFILE'));
 
 		echo '<p>&nbsp;'._('Happy learning...').'</p>';
 	break;