+++ CHANGES.md
@@ -8,6 +8,7 @@ Changes in 3.4.1
 - #191 Fix PHP notices thanks to @vanyog in GetStaffList.fnc.php, GetStuList.fnc.php & ListOutput.fnc.php
 - Set current SchoolYear on login in index.php, Side.php & Portal.php
 - Fix SQL error when UserSchool() not set in Config.fnc.php
+- Optimize: remove SCHOOL_DATE index (events) & group LO options in vars in Portal.php
 
 Changes in 3.4
 --------------